requirements.txt
sentence-transformers==2.2.2
faiss-cpu==1.8.0
numpy==1.26.0

